+++ Sandbox/Source/Systems/Rendering/IBLSkybox.cpp
@@ -145,8 +145,8 @@ void IBLSkybox::GenBuffers(unsigned int texture, unsigned int size)
 	for (unsigned int mip = 0; mip < maxMipLevels; ++mip)
 	{
 		// reisze framebuffer according to mip-level size.
-		unsigned int mipWidth = 128 * std::pow(0.5, mip);
-		unsigned int mipHeight = 128 * std::pow(0.5, mip);
+		unsigned int mipWidth = 128 * static_cast<unsigned int>(std::pow(0.5, mip));
+		unsigned int mipHeight = 128 * static_cast<unsigned int>(std::pow(0.5, mip));
 		glBindRenderbuffer(GL_RENDERBUFFER, captureRBO);
 		glRenderbufferStorage(GL_RENDERBUFFER, GL_DEPTH_COMPONENT24, mipWidth, mipHeight);
 		glViewport(0, 0, mipWidth, mipHeight);
